deps/decomposition: rename phono to pipe and move pipe to the top
