+++ pipe_internal_test.go
