+++ pipe_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/pipelined/mock"
-	"github.com/pipelined/phono/pipe"
+	"github.com/pipelined/pipe"
 	"go.uber.org/goleak"
 )
